+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -3,6 +3,8 @@
 Release History
 ===============
 
+* add listen-address option to "az aks port-forward"
+
 2.3.17
 ++++++
 * az aks enable-addons /disable-addons: support case insensitive name